requirements.txt
Flask>=2.0.0
Flask-CORS>=4.0.0
requests>=2.25.0
pandas>=1.3.0
scipy>=1.7.0
python-dateutil>=2.8.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
werkzeug>=2.0.0
flask-talisman>=1.0.0
redis>=4.0.0
celery>=5.2.0 